// src/interfaces/GameObject.d.ts
// Generated by https://quicktype.io
//
// To change quicktype's target language, run command:
//
//   "Set quicktype target language"

export interface GameObject {
    ObjectTypeId:                 string;
    DisplayInfo:                  DisplayInfo;
    Categories:                   Category[];
    Image:                        TmpImage | {};
    StandardSupplyCost:           number | null;
    StandardPopulationCost:       number | null;
    StandardEnergyCost:           number | null;
    EffectivenessAgainstInfantry: string | null;
    EffectivenessAgainstVehicles: string | null;
    EffectivenessAgainstAir:      string | null;
}

export interface Category {
    Id:        number;
    Type:      string;
    Identity:  string;
    Autoroute: string;
}

export interface DisplayInfo {
    Id:   number;
    Type: string;
    View: DisplayInfoView;
}

export interface DisplayInfoView {
    Status:               string;
    Common:               Common;
    Identity:             string;
    BatchLocalization:    BatchLocalization;
    Localization:         Localization;
    Title:                string;
    HW2ObjectDisplayInfo: HW2ObjectDisplayInfo;
}

export interface BatchLocalization {
    IsLocked: boolean;
}

export interface Common {
    Owner:        string;
    CreatedUtc:   string;
    ModifiedUtc:  string;
    PublishedUtc: string;
    Container:    number;
}

export interface HW2ObjectDisplayInfo {
    Name:        string;
    Description: string;
}

export interface Localization {
    Culture:             string;
    MasterContentItemId: string;
}

export interface TmpImage {
    Id:   number;
    Type: string;
    View: ImageView;
}

export interface ImageView {
    Status:   string;
    Common:   Common;
    Media:    Media;
    Image:    ViewImage;
    Title:    string;
    Identity: string;
}

export interface ViewImage {
    Width:  number;
    Height: number;
}

export interface Media {
    MediaUrl:      string;
    MimeType:      string;
    Caption:       string;
    AlternateText: string;
    FolderPath:    string;
    FileName:      string;
}